Move mock to own file
