+++ frame/vesting/src/tests.rs
@@ -1,124 +1,9 @@
-use super::*;
-use crate as pallet_vesting;
-
-use frame_support::{assert_ok, assert_noop, parameter_types};
-use sp_core::H256;
-use sp_runtime::{
-	testing::Header,
-	traits::{BlakeTwo256, IdentityLookup, Identity, BadOrigin},
-};
+use frame_support::{assert_noop, assert_ok};
 use frame_system::RawOrigin;
+use sp_runtime::traits::BadOrigin;
 
-type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-type Block = frame_system::mocking::MockBlock<Test>;
-
-frame_support::construct_runtime!(
-	pub enum Test where
-		Block = Block,
-		NodeBlock = Block,
-		UncheckedExtrinsic = UncheckedExtrinsic,
-	{
-		System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
-		Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
-		Vesting: pallet_vesting::{Pallet, Call, Storage, Event<T>, Config<T>},
-	}
-);
-
-parameter_types! {
-	pub const BlockHashCount: u64 = 250;
-	pub BlockWeights: frame_system::limits::BlockWeights =
-		frame_system::limits::BlockWeights::simple_max(1024);
-}
-impl frame_system::Config for Test {
-	type BaseCallFilter = ();
-	type BlockWeights = ();
-	type BlockLength = ();
-	type DbWeight = ();
-	type Origin = Origin;
-	type Index = u64;
-	type BlockNumber = u64;
-	type Hash = H256;
-	type Call = Call;
-	type Hashing = BlakeTwo256;
-	type AccountId = u64;
-	type Lookup = IdentityLookup<Self::AccountId>;
-	type Header = Header;
-	type Event = Event;
-	type BlockHashCount = BlockHashCount;
-	type Version = ();
-	type PalletInfo = PalletInfo;
-	type AccountData = pallet_balances::AccountData<u64>;
-	type OnNewAccount = ();
-	type OnKilledAccount = ();
-	type SystemWeightInfo = ();
-	type SS58Prefix = ();
-	type OnSetCode = ();
-}
-parameter_types! {
-	pub const MaxLocks: u32 = 10;
-}
-impl pallet_balances::Config for Test {
-	type Balance = u64;
-	type DustRemoval = ();
-	type Event = Event;
-	type ExistentialDeposit = ExistentialDeposit;
-	type AccountStore = System;
-	type MaxLocks = MaxLocks;
-	type MaxReserves = ();
-	type ReserveIdentifier = [u8; 8];
-	type WeightInfo = ();
-}
-parameter_types! {
-	pub const MinVestedTransfer: u64 = 256 * 2;
-	pub static ExistentialDeposit: u64 = 0;
-}
-impl Config for Test {
-	type Event = Event;
-	type Currency = Balances;
-	type BlockNumberToBalance = Identity;
-	type MinVestedTransfer = MinVestedTransfer;
-	type WeightInfo = ();
-}
-
-pub struct ExtBuilder {
-	existential_deposit: u64,
-}
-impl Default for ExtBuilder {
-	fn default() -> Self {
-		Self {
-			existential_deposit: 1,
-		}
-	}
-}
-impl ExtBuilder {
-	pub fn existential_deposit(mut self, existential_deposit: u64) -> Self {
-		self.existential_deposit = existential_deposit;
-		self
-	}
-	pub fn build(self) -> sp_io::TestExternalities {
-		EXISTENTIAL_DEPOSIT.with(|v| *v.borrow_mut() = self.existential_deposit);
-		let mut t = frame_system::GenesisConfig::default().build_storage::<Test>().unwrap();
-		pallet_balances::GenesisConfig::<Test> {
-			balances: vec![
-				(1, 10 * self.existential_deposit),
-				(2, 20 * self.existential_deposit),
-				(3, 30 * self.existential_deposit),
-				(4, 40 * self.existential_deposit),
-				(12, 10 * self.existential_deposit)
-			],
-		}.assimilate_storage(&mut t).unwrap();
-		pallet_vesting::GenesisConfig::<Test> {
-			vesting: vec![
-				(1, 0, 10, 5 * self.existential_deposit),
-				(2, 10, 20, 0),
-				(12, 10, 20, 5 * self.existential_deposit)
-			],
-		}.assimilate_storage(&mut t).unwrap();
-		let mut ext = sp_io::TestExternalities::new(t);
-		ext.execute_with(|| System::set_block_number(1));
-		ext
-	}
-}
+use super::*;
+use crate::mock::{Balances, ExtBuilder, System, Test, Vesting};
 
 #[test]
 fn check_vesting_status() {
